ingestors/pyproject.toml
[project]
name = "ingestors"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncua>=1.1.8",
    "pandas>=3.0.1",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
]
